# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  timestamp?: string | null;
}

export interface User {
  id: number;
  name: string;
  role: string;
}

export interface ActivityEntry {
  timestamp: string;
  action: string;
  detail: string;
}

export interface UserDetail {
  id: number;
  name: string;
  role: string;
  email: string;
  /** @nullable */
  lastLogin?: string | null;
  createdAt: string;
  activityLog: ActivityEntry[];
}

export interface CreateUserBody {
  name: string;
  role: string;
  email: string;
}

export interface UpdateUserBody {
  name?: string;
  role?: string;
  email?: string;
}

export type UserList = User[];

export interface RequestLog {
  id: string;
  timestamp: string;
  method: string;
  route: string;
  statusCode: number;
  latencyMs: number;
  userAgent?: string;
}

export type RequestLogList = RequestLog[];

export interface Deployment {
  id: string;
  version: string;
  timestamp: string;
  status: string;
  environment: string;
  pipeline: string;
  commitHash?: string;
  author?: string;
  duration?: number;
}

export type DeploymentList = Deployment[];

export interface Pod {
  name: string;
  status: string;
  restarts: number;
  cpuPercent: number;
  memoryMb: number;
  node: string;
  age: string;
  namespace: string;
}

export type PodList = Pod[];

export interface Incident {
  id: string;
  title: string;
  severity: string;
  status: string;
  startedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  rootCause?: string | null;
}

export type IncidentList = Incident[];

export interface AlertConfig {
  errorRateWarning: number;
  errorRateCritical: number;
  latencyP95Warning: number;
  latencyP95Critical: number;
  podRestartWarning: number;
  podRestartCritical: number;
}