# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
      required:
        - id
        - name
        - role
    UserDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        email:
          type: string
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
        activityLog:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
      required:
        - id
        - name
        - role
        - email
        - createdAt
        - activityLog
    ActivityEntry:
      type: object
      properties:
        timestamp:
          type: string
        action:
          type: string
        detail:
          type: string
      required:
        - timestamp
        - action
        - detail
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
      required:
        - name
        - role
        - email
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/User"
...