# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
  service: zod.string().nullish(),
  timestamp: zod.string().nullish(),
});

/**
 * @summary List users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a user
 */
export const CreateUserBody = zod.object({
  name: zod.string(),
  role: zod.string(),
  email: zod.string(),
});

/**
 * @summary Get user details
 */
export const GetUserParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
  email: zod.string(),
  lastLogin: zod.string().nullish(),
  createdAt: zod.string(),
  activityLog: zod.array(
    zod.object({
      timestamp: zod.string(),
      action: zod.string(),
      detail: zod.string(),
    }),
  ),
});

/**
 * @summary Update a user
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  name: zod.string().optional(),
  role: zod.string().optional(),
  email: zod.string().optional(),
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
});

/**
 * @summary Delete a user
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get recent request logs
 */
export const GetRequestLogsResponseItem = zod.object({
  id: zod.string(),
  timestamp: zod.string(),
  method: zod.string(),
  route: zod.string(),
  statusCode: zod.number(),
  latencyMs: zod.number(),
  userAgent: zod.string().optional(),
});
export const GetRequestLogsResponse = zod.array(GetRequestLogsResponseItem);

/**
 * @summary Get deployment history
 */
export const GetDeploymentsResponseItem = zod.object({
  id: zod.string(),
  version: zod.string(),
  timestamp: zod.string(),
  status: zod.string(),
  environment: zod.string(),
  pipeline: zod.string(),
  commitHash: zod.string().optional(),
  author: zod.string().optional(),
  duration: zod.number().optional(),
});
export const GetDeploymentsResponse = zod.array(GetDeploymentsResponseItem);

/**
 * @summary Get pod/node health
 */
export const GetPodsResponseItem = zod.object({
  name: zod.string(),
  status: zod.string(),
  restarts: zod.number(),
  cpuPercent: zod.number(),
  memoryMb: zod.number(),
  node: zod.string(),
  age: zod.string(),
  namespace: zod.string(),
});
export const GetPodsResponse = zod.array(GetPodsResponseItem);

/**
 * @summary Get incident timeline
 */
export const GetIncidentsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  severity: zod.string(),
  status: zod.string(),
  startedAt: zod.string(),
  resolvedAt: zod.string().nullish(),
  durationMinutes: zod.number().nullish(),
  rootCause: zod.string().nullish(),
});
export const GetIncidentsResponse = zod.array(GetIncidentsResponseItem);

/**
 * @summary Get alert thresholds
 */
export const GetAlertConfigResponse = zod.object({
  errorRateWarning: zod.number(),
  errorRateCritical: zod.number(),
  latencyP95Warning: zod.number(),
  latencyP95Critical: zod.number(),
  podRestartWarning: zod.number(),
  podRestartCritical: zod.number(),
});

/**
 * @summary Update alert thresholds
 */
export const UpdateAlertConfigBody = zod.object({
  errorRateWarning: zod.number(),
  errorRateCritical: zod.number(),
  latencyP95Warning: zod.number(),
  latencyP95Critical: zod.number(),
  podRestartWarning: zod.number(),
  podRestartCritical: zod.number(),
});

export const UpdateAlertConfigResponse = zod.object({
  errorRateWarning: zod.number(),
  errorRateCritical: zod.number(),
  latencyP95Warning: zod.number(),
  latencyP95Critical: zod.number(),
  podRestartWarning: zod.number(),
  podRestartCritical: zod.number(),
});