# lib/api-zod/src/generated/types/userDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntry } from "./activityEntry";

export interface UserDetail {
  id: number;
  name: string;
  role: string;
  email: string;
  /** @nullable */
  lastLogin?: string | null;
  createdAt: string;
  activityLog: ActivityEntry[];
}